test: cover web auth session fallback branches
